+++ optd-datafusion/src/main.rs
@@ -76,3 +76,13 @@ async fn main() -> Result<()> {
     }
     Ok(())
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[tokio::test]
+    async fn test_main() {
+    }
+}
+
